Fix of gcc warnings.

+++ src/ch1/ch1/ch1.cpp
@@ -61,7 +61,7 @@ void ex1_3_37(int32_t n, int32_t m)
     {
       const auto re{queue.remove(indexToDel)};
       fmt::print("{} ", re.value_or(-1));
-      indexToDel = (indexToDel + m - 1) % queue.size();
+      indexToDel = (indexToDel + m - 1) % static_cast<int32_t>(queue.size());
     }
   }
   const auto re{queue.remove(indexToDel)};
